Create findandreplace.html
Let’s test the JQuery selectors and manipulation commands you learned this week. Download this fileLinks to an external site., which has True Facts about Hedgehogs, and automated tests in the upper-right corner.

Save the file as a plain HTML file and then open in your text editor.

Below the HTML content, you’ll see a script block with partial code you’ll need to complete so that each pre-defined text passes (turns green).

Assign the desired value to variables already listed,
Where there is no a variable, make the specified page changes to pass that test.
The tests are not always perfect, so let me know if you think you've got a correct answer but you can't get it to turn green.